+++ data-raw/another.test.R
@@ -1,9 +1,10 @@
 library(BSgenome.Hsapiens.1000genomes.hs37d5)
 
 spect <- 
-  ICAMS::MutectVCFFilesToCatalog(
+  ICAMS::VCFsToCatalogs(
     "data-raw/mSigAct-example-VCFs/mSigAct-example-VCFs/Mutect-mSigAct-example-GRCh37-s1.vcf",
-    ref.genome = BSgenome.Hsapiens.1000genomes.hs37d5)
+    ref.genome = BSgenome.Hsapiens.1000genomes.hs37d5,
+    variant.caller = "mutect")
 
 
 # incomplete