+++ tests/testthat/test_MAPAssignActivity.R
@@ -27,8 +27,6 @@ test_that("MAPAssignActivity for ID Catalog", {
     mc.cores.per.sample     = 30,
     seed                    = 8787)
   
-  expect_equal(length(retval), 4)
-  
   expect_true (all.equal(retval$proposed.assignment[, 1],
                          round(
                          c(ID1 = 60.0957928396838, 
@@ -99,8 +97,6 @@ test_that("Use sparse assignment for MAPAssignActivity", {
     mc.cores.per.sample     = 30,
     seed                    = 2351,
     use.sparse.assign       = TRUE)
-  expect_equal(nrow(retval2$proposed.assignment), 14)
-  
   
   retval3 <- SparseAssignActivity(spectra = catalog,
                                   sigs = sigs,