+++ tests/testthat/test_multinomial_likelihood.R
@@ -22,7 +22,7 @@ test_that("Use multinomial distribution to calculate likelihood", {
                       output.dir = file.path(tempdir(), "skin.mapout1"))
   
   expect_equal(MAP.out1$reconstruction.distances$MAP.distances$cosine,
-               0.9525457, tolerance = 1e-5)
+               0.9517956, tolerance = 1e-2)
   
   # Use multinomial distribution to calculate log likelihood (the default)
   MAP.out2 <-